Create FUNDING.yml
I have been informally educating myself for three years on most every aspect of law- mainly [Ohio] state and federal. I'm cognitively disabled by the Marine Corps base Camp Lejeune toxic water scandal. I'm a victim of gang stalking at the local and state govt level and a fugitive from injustice at this time, but I'm determined to beat the pants off of the so-called "JUSTICE" system - I already forced the court to provide me a remote hearing and got one felony warrant recalled without having to turn myself in. I'm a firm believer in Our human rights being impossible to remove and nOW I have gained the knowledge of enough of the law to force the powers that be to hear meOW! (8 I'm even a homeless veteran - almost three years nOW! I've mastered the art of legal research and I love to help others! I have two pretty incredible app ideas with huge potential to be both lucrative and impactful AF for most ppL actually. Who wants to go down in history as someone who actually assisted me in my endeavors‽ I'm already beginning to strike back at the government representatives that have criminally me. Wilfully depriving someone in the US of their LEGAL RIGHTS under color of law, or government, is a misdemeanor of the 1st degree. If the victim is a member of a vulnerable, or protected, class the crime is at least a felony five (or the lowest felony degree). And if two or more colluded to willfully deprive someone in the United States of one or more of their legal rights under color of law the crime is much more serious! I'm grateful for anything anyone can do to help me out with this fight. I'm willing to die for actual JUSTICE. Steve Jobs said it best – "[Those] "crazy enough to think that they can change the world are the ones who dO!" (8